requirements.txt
npm==10.8.2
axios==1.7.9
fastapi==0.115.6
matplotlib==3.9.2
mido==1.3.3
numpy==2.0.2
pillow==11.0.0
